requirements.txt
asgiref==3.5.2
boto3==1.24.80
botocore==1.27.80
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-import-export==3.0.1
django-material-admin==1.8.6
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23
urllib3==1.26.12
uWSGI==2.0.20
